Netology. Spring Core. Spring Web MVC. Task 1. Project's files uploaded.

+++ src/main/java/com/gmail/at/kotamadeo/service/PostService.java
@@ -4,16 +4,14 @@ import com.gmail.at.kotamadeo.exception.NotFoundException;
 import com.gmail.at.kotamadeo.model.Post;
 import com.gmail.at.kotamadeo.repository.PostRepository;
 import lombok.RequiredArgsConstructor;
-import lombok.experimental.FieldDefaults;
+import org.springframework.stereotype.Service;
 
 import java.util.List;
 
-import static lombok.AccessLevel.PRIVATE;
-
 @RequiredArgsConstructor
-@FieldDefaults(level = PRIVATE)
+@Service
 public class PostService {
-    final PostRepository repository;
+    private final PostRepository repository;
 
     public List<Post> all() {
         return repository.all();
@@ -28,6 +26,9 @@ public class PostService {
     }
 
     public void removeById(long id) {
+        if (repository.getById(id).isEmpty()) {
+            throw new NotFoundException();
+        }
         repository.removeById(id);
     }
 }
